+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Added
 - Add ``BackgroundTask`` model to enable support for observable long running
   backend tasks
 - Add ``BackgroundTaskConsumer`` class to handle starting background jobs
+- Make ``Kubernetes`` overcommitment factors configurable via settings
 
 Changed
 -------